Security update wip commit, content service anon services
Updated authentication by separating out providers and created
a provider model for loading additional passport strategies
Anon strategy updated to allow for ANON role check on anonymous
connection.
Added baseline implementation for following providers and are a work in progress.
- Facebook
- GitHub
- Google
- LinkedIn
- Twitter
Adding the context provider to the request for use with other application requests.